requirements.txt
pandas
sentence-transformers
umap-learn
requests
